tests: Relocate GetHighestCPUNumberAmongNodes to libnode
